+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## Unrelease
+
+**Release data:** `N/A`
+
+### Changed
+
+- Renamed `BitField.has` to `BitField.contains`.
+
 ## 1.0.0
 
 **Release date:** `2023-11-29`